Additional fixes around logging
