+++ Dockerfile
@@ -118,6 +118,7 @@ RUN apt install nodejs
 ENV SERVING_PORT="8082"
 ENV SERVING_HOST="0.0.0.0"
 ENV ETCD_ARCH=""
+ENV MEILI_LOG_LEVEL="WARN"
 
 EXPOSE 7878
 EXPOSE 80
