+++ client/src/featureform/client.py
@@ -32,7 +32,13 @@ class Client(ResourceClient, ServingClient):
     """
 
     def __init__(
-        self, host=None, local=False, insecure=False, cert_path=None, dry_run=False
+        self,
+        host=None,
+        local=False,
+        insecure=False,
+        cert_path=None,
+        dry_run=False,
+        debug=False,
     ):
         ResourceClient.__init__(
             self,
@@ -41,12 +47,18 @@ class Client(ResourceClient, ServingClient):
             insecure=insecure,
             cert_path=cert_path,
             dry_run=dry_run,
+            debug=debug,
         )
         # Given both ResourceClient and ServingClient are instantiated together, if dry_run is True, then
         # the ServingClient cannot be instantiated due to a conflict the local and host arguments.
         if not dry_run:
             ServingClient.__init__(
-                self, host=host, local=local, insecure=insecure, cert_path=cert_path
+                self,
+                host=host,
+                local=local,
+                insecure=insecure,
+                cert_path=cert_path,
+                debug=debug,
             )
 
     def dataframe(
