+++ metadata/etcd.go
@@ -311,7 +311,7 @@ func (lookup EtcdResourceLookup) serializeResource(res Resource) ([]byte, error)
 func (lookup EtcdResourceLookup) deserialize(value []byte) (EtcdRow, error) {
 	var tmp EtcdRowTemp
 	if err := json.Unmarshal(value, &tmp); err != nil {
-		return EtcdRow{}, errors.Wrap(err, fmt.Sprintf("failed To Parse Resource: %s", value))
+		return EtcdRow{}, errors.Wrap(err, fmt.Sprintf("failed To parse resource: %s", value))
 	}
 	msg := EtcdRow{
 		ResourceType: ResourceType(tmp.ResourceType),
@@ -323,22 +323,21 @@ func (lookup EtcdResourceLookup) deserialize(value []byte) (EtcdRow, error) {
 
 func (lookup EtcdResourceLookup) Lookup(id ResourceID) (Resource, error) {
 	key := createKey(id)
-	fmt.Printf("Lookup Key: %s\n", key)
 	resp, err := lookup.Connection.Get(key)
 	if err != nil || len(resp) == 0 {
 		return nil, &ResourceNotFound{id, err}
 	}
 	msg, err := lookup.deserialize(resp)
 	if err != nil {
-		return nil, errors.Wrap(err, fmt.Sprintf("lookup deserialize: %s", id))
+		return nil, errors.Wrap(err, fmt.Sprintf("failed to deserialize: %s", id))
 	}
 	resType, err := lookup.createEmptyResource(msg.ResourceType)
 	if err != nil {
-		return nil, errors.Wrap(err, fmt.Sprintf("lookup create: %s", id))
+		return nil, errors.Wrap(err, fmt.Sprintf("failed to create empty resource: %s", id))
 	}
 	resource, err := lookup.Connection.ParseResource(msg, resType)
 	if err != nil {
-		return nil, errors.Wrap(err, fmt.Sprintf("lookup parse: %s", id))
+		return nil, errors.Wrap(err, fmt.Sprintf("failed to parse resource: %s", id))
 	}
 	return resource, nil
 }
@@ -431,11 +430,11 @@ func (lookup EtcdResourceLookup) Submap(ids []ResourceID) (ResourceLookup, error
 
 	for _, id := range ids {
 		key := createKey(id)
-		value, err := lookup.Connection.Get(key)
-		if err != nil {
+		resp, err := lookup.Connection.Get(key)
+		if err != nil || len(resp) == 0 {
 			return nil, &ResourceNotFound{id, err}
 		}
-		etcdStore, err := lookup.deserialize(value)
+		etcdStore, err := lookup.deserialize(resp)
 		if err != nil {
 			return nil, errors.Wrap(err, fmt.Sprintf("submap deserialize: %s", id))
 		}
