+++ tests/end_to_end/features/cloud_credentials.feature
@@ -3,7 +3,7 @@ Feature: Credentials
   Scenario Outline: Create AWS Credentials
     Given Featureform is installed
     When I create the AWS credentials with an access key "<access_key>" and secret key "<secret_key>"
-    Then An exception "<exception>" should be raised
+    Then An exception that "matches" "<exception>" should be raised
     Examples:
       | access_key | secret_key | exception |
       | AXASCASCA  | asdvadvadv | None  |
@@ -14,7 +14,7 @@ Feature: Credentials
   Scenario Outline: Create GCP Credentials
     Given Featureform is installed
     When I create the GCP credentials with a project id "<project_id>" and credentials path "<credentials_path>"
-    Then An exception "<exception>" should be raised
+    Then An exception that "matches" "<exception>" should be raised
     Examples:
       | project_id | credentials_path | exception |
       | my-project | test_files/dummy_creds.json | None  |
